Update to 3.5.0
**Backend Rework Edition**
This update has been a long time in the making. I made the first addition to the change log almost a month ago, but I was waiting on user facing releases to drop the patch notes. Most of the code edits this time were backend reworking, as you can tell from some of the items in the list. The bot's a lot more robust and easy to work on now, and I paid down a lot of technical debt. Enjoy.

User Features
- Added the !youtube command to search youtube for a video given its title as the argument
- Reworked subreddit detection to not only be more reliable, but it also supports multiple subreddits per message and provides more information about each community
- Added a "forever alone" tracker for Colton. I hope he enjoys it
- Removed the !jpeg command because I realized it used a blocking library for url requests, and it was more effort to get it working asynchronously than it was worth
- !list title no longer accepts a list ID as a parameter because no one has or ever will use that feature and it makes syntax more consistent

Bug Fixes and Text Changes
- Fixed a bug where !list title would fail
- Fixed an inaccuracy in !bestGirl's User Guide
- Fixed an inaccuracy in !list's User Guide

Dev tools
- Improved error reporting to make locating bugs easier, which will speed up bug fixes
- Rebuilt several dev-only commands that only I know about
- Added feature where I can now change the bot's "currently playing" status from within Discord
- Deleted some unused code to make the bot more GitHub worthy