v4.2.0
Updated the repository to the current version of the bot. This version does not have specific changes but is rather the accumulation of small bug fixes and changes over time. The most drastic change is the inclusion of the scheduler, a subsystem for assigning the bot to perform specific events at specific reoccurring intervals