Created StringCache
This is very similar to the system previously used, it's just generalized out so it can be reused for all sorts of things. It reduces stress on API calls by caching a list of strings and then refilling it when the list gets low. The process for retrieving strings (in `!meow`'s case, cat image urls) is defined by subclassing StringCache and defining the `async gather()` method.